Created Login and Register Controllers.
