--- src/main/java/brainyBunch/liftoffgroup3/controller/HomeController.java
@@ -1,16 +1,0 @@
-package brainyBunch.liftoffgroup3.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-
-@Controller
-@RequestMapping("/")
-public class HomeController {
-
-    @GetMapping("")
-    public String viewHomePage() {
-        return "index";
-    }
-}
