--- src/main/java/brainyBunch/liftoffgroup3/controller/RegisterController.java
@@ -1,18 +1,0 @@
-package brainyBunch.liftoffgroup3.controller;
-
-import brainyBunch.liftoffgroup3.model.User;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class RegisterController {
-
-    @GetMapping("/register")
-    public String viewRegistrationForm(Model model){
-        // create model object to store form data
-        User user = new User();
-        model.addAttribute("user", user);
-        return "register";
-    }
-}
